refactor(radiant): reorganize server SSR layout and centralize attribute policy
