+++ packages/radiant/src/server/install-light-dom-shim.ts
@@ -1,0 +1,1 @@
+import './install/install-light-dom-shim';
