+++ packages/radiant/src/server/light-dom-shim.ts
@@ -1,0 +1,1 @@
+export * from './shim/light-dom-shim';
