+++ packages/radiant/src/server/radiant-element-ssr-bridge.ts
@@ -1,3 +1,7 @@
+/**
+ * @deprecated Import `@ecopages/radiant/server/radiant-element-ssr` instead.
+ * Kept for compatibility with tooling that still resolves this entry path.
+ */
 export {
 	createRadiantElementSsrService,
 	getOrCreateRadiantElementSsrRuntime,
