+++ packages/radiant/src/server/render-component.ts
@@ -1,0 +1,1 @@
+export * from './adapters/render-component';
