+++ packages/radiant/src/server/shim/minimal-dom/align-host-tag-name.ts
@@ -1,4 +1,4 @@
-import { getCustomElementTagName } from '../../core/custom-element-metadata';
+import { getCustomElementTagName } from '../../../core/custom-element-metadata';
 
 function isMinimalDomElement(host: object): host is { localName: string; tagName: string } {
 	return (
